update on admin and art page

+++ adminapp/views.py
@@ -1,14 +1,17 @@
 from django.shortcuts import render,redirect
 from django.views import View
 # Create your views here.
-from .forms import CategoryForm,SubcatsForm,RegisterForm,ProductForm
+from .forms import CategoryForm,SubcatsForm,ProductForm
 from . models import *
 from django.shortcuts import render, get_object_or_404, redirect
-
-
 from django.contrib.auth import logout
+
+
 class adminaddcat(View):
     def get(self,request):
+        if not request.user.is_superuser:
+            return redirect('/')
+            
         categoryform=CategoryForm()
         context={'categoryform':categoryform,'title':'Add category'}
         return render(request,'pages/adminaddcat.html',context)
@@ -21,62 +24,68 @@ class adminaddcat(View):
         else:
               return redirect('adminapp:adminaddcat')
         
-class viewcategory(View):
-    def get(self,request):
-        category=Category.objects.all()
-        context={"category":category}
-        return render(request,'pages/adminviewcategory.html',context)
+
 
 class updatecategory(View):
     def get(self,request,pk):
+        if not request.user.is_superuser:
+            return redirect(request.META.get('HTTP_REFERER','/'))
         get_cart_by_id=get_object_or_404(Category,pk=pk)
         form=CategoryForm(instance=get_cart_by_id)
         context={"form":form,"get_cart_by_id":get_cart_by_id}
         return render(request,'pages/update_category.html',context)
     def post(self,request,pk):
+        if not request.user.is_superuser:
+            return redirect(request.META.get('HTTP_REFERER','/'))
         
         get_cart_by_id=get_object_or_404(Category,pk=pk)
         form=CategoryForm(request.POST,instance=get_cart_by_id)
         if form.is_valid():
             form.save()
-            return redirect("adminapp:viewcategory")
+            return redirect("adminapp:adminaddcat")
         else:
             return redirect(request.META.get('HTTP_REFERER','/'))
         
 class deletecategory(View):
     def get(self,request,pk):
+        if not request.user.is_superuser:
+            return redirect(request.META.get('HTTP_REFERER','/'))
         get_object_or_404(Category,pk=pk).delete()
         return redirect(request.META.get('HTTP_REFERER','/'))
     
-# adding subcategory into database
-class adminsubcats(View):
-    def get(self,request):
-        subform=SubcatsForm()
-        context={'subform':subform}
-        return render(request,'pages/adminsubcats.html',context)
 
+class subcategory(View):
+    def get(self,request):
+        if not request.user.is_superuser:
+            return redirect(request.META.get('HTTP_REFERER','/'))
+        subform=SubcatsForm()
+        getsubcategory=Subcategory.objects.all()
+        context={"subcat":getsubcategory,'subform':subform}
+        return render(request,'pages/viewsubcategory.html',context)
+    
     def post(self, request):
+        if not request.user.is_superuser:
+            return redirect(request.META.get('HTTP_REFERER','/'))
         subform=SubcatsForm(request.POST)
         if subform.is_valid():
             subform.save()
-            return redirect('adminapp:adminsubcats')
+            return redirect('adminapp:viewsubcategory')
         else:
-            return redirect('adminapp:adminsubcats')
+            return redirect('adminapp:viewsubcategory')
         
-class viewsubcategory(View):
-    def get(self,request):
-        getsubcategory=Subcategory.objects.all()
-        context={"subcat":getsubcategory}
-        return render(request,'pages/viewsubcategory.html',context)
-    
+
 class updatesubcategory(View):
     def get(self,request,pk):
+        if not request.user.is_superuser:
+            return redirect(request.META.get('HTTP_REFERER','/'))
         subcategory=get_object_or_404(Subcategory,pk=pk)
         form=SubcatsForm(instance=subcategory)
         context={"form":form,'subcategory':subcategory}
         return render(request,'pages/update_subcategory.html',context)
     
     def post(self,request,pk):
+        if not request.user.is_superuser:
+            return redirect(request.META.get('HTTP_REFERER','/'))
         subcategory=get_object_or_404(Subcategory,pk=pk)
         form=SubcatsForm(request.POST,instance=subcategory)
         if form.is_valid():
@@ -87,31 +96,14 @@ class updatesubcategory(View):
          
 class deletesubcategory(View):
     def get(self,request,pk):
+        if not request.user.is_superuser:
+            return redirect(request.META.get('HTTP_REFERER','/'))
         get_object_or_404(Subcategory,pk=pk).delete()
         return redirect(request.META.get('HTTP_REFERER','/'))
     
     
         
-# registration forms
-class registeruser(View):
-    def get(self,request):
-        registerform=RegisterForm()
-        context={'registerform':registerform,'title':'registration'}
-        return render(request,'pages/registeruser.html',context)
-    def post(self,request):
-        registerform=RegisterForm(request.POST)
-        if registerform.is_valid():
-            registerform.save()
-            return redirect('adminapp:login')
-        else:
-            return redirect('adminapp:registeruser')
-        
 
-# logout
-
-def logoutuser(request):
-    logout(request)
-    return redirect('adminapp:login')
